pyproject.toml
[tool.poetry]
name = "gitlab-review-analyzer"
version = "0.1.0"
description = "GitLab MR review sentiment analysis tool"
authors = ["User"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12.2"
python-gitlab = "^6.0.0"
pandas = "^2.3.0"
numpy = "^2.3.0"
matplotlib = "^3.10.3"
seaborn = "^0.13.2"
jinja2 = "^3.1.6"
textblob = "^0.19.0"
vadersentiment = "^3.3.2"
ollama = "^0.5.1"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
